Consolidate eval
