+++ docs/api.md
@@ -3,5 +3,3 @@
 ::: iup
 
 ::: iup.models
-
-::: iup.eval
